backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pymupdf==1.24.3
pytesseract==0.3.10
Pillow==10.3.0
httpx==0.27.0
reportlab==4.2.0
python-dateutil==2.9.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.7.1
pydantic-settings==2.2.1
stripe==9.9.0
resend==2.0.0
